requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
openai
numpy